% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygentank-package.R
\docType{package}
\name{roxygentank-package}
\alias{roxygentank}
\alias{roxygentank-package}
\title{roxygentank: Templates for documentation of R files.}
\description{
Creates templates for documentation of R files. Contains comprehensive documentation of roxygen2 and formatting tags, and also suggested minimal templates.
}
\author{
\strong{Maintainer}: Robert Amezquita \email{robert.amezquita@fredhutch.org}

}
\keyword{internal}
