% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygentank.R
\name{roxygentank_templates}
\alias{roxygentank_templates}
\alias{roxygentank_summary}
\alias{roxygentank_minimal}
\title{Create templates for new R functions}
\usage{
roxygentank_summary(name = NULL)

roxygentank_minimal(name = NULL)
}
\arguments{
\item{name}{File name, without extension; will create if it doesn't
already exist; if left null, creates a new file \code{R/roxygentank_template.R}}
}
\value{
none; called for its side-effect of creating an R file
}
\description{
Creates a .R file with all available roxygen documentation tags
with descriptions of each one. Places the new .R file within the current
project's \code{R/} directory. \code{\link[=roxygentank_summary]{roxygentank_summary()}} creates a new file
with documentation of select tags and formatting options, while
\code{\link[=roxygentank_minimal]{roxygentank_minimal()}} only provides a basic .R with some
suggested tags and organization.
}
